+++ src/NServiceBus.Core/Recoverability/SecondLevelRetries/ImmediateRetry.cs
@@ -1,0 +1,6 @@
+﻿namespace NServiceBus
+{
+    class ImmediateRetry : RecoveryAction
+    {
+    }
+}
