+++ src/NServiceBus.Core/Recoverability/SecondLevelRetries/MoveToErrorQueue.cs
@@ -1,0 +1,6 @@
+﻿namespace NServiceBus
+{
+    class MoveToErrorQueue : RecoveryAction
+    {
+    }
+}
