+++ src/NServiceBus.Core/Recoverability/SecondLevelRetries/RecoveryAction.cs
@@ -1,0 +1,6 @@
+﻿namespace NServiceBus
+{
+    abstract class RecoveryAction
+    {
+    }
+}
